innoai-backend-main/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Limiter==2.6.2
Flask-Caching==2.0.1
Flask-SQLAlchemy==3.0.5
redis==4.5.5
black==23.7.0
pytest==7.4.0
pydantic==1.10.7
gunicorn==21.2.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
Werkzeug==2.3.7
tensorflow==2.15.0
tensorflow-hub==0.13.0
tensorflow-estimator
tensorboard
pinecone-client==2.2.2
pdfminer.six==20201018
langchain
openai==1.0.0
langchain-core
protobuf==3.20.3
PyMuPDF
colorama==0.4.6
spacy
tf-keras==2.15
nltk==3.7
chromadb==0.5.0
typer
pymongo
